requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
<<<<<<< HEAD
asyncpg==0.30.*
httpx==0.28.*
sentence-transformers==3.3.*
lxml==5.3.*
pydantic-settings==2.7.*
sse-starlette==2.2.*
google-generativeai==0.8.*
python-dotenv==1.0.*
tqdm==4.67.*
=======
httpx==0.28.*
pydantic-settings==2.7.*
python-dotenv==1.0.*
>>>>>>> origin/elevenlabs
